improved work with cell comments in XSSF, also added support for cell comments to SS interfaces
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@884918 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ss/usermodel/Comment.java
@@ -59,7 +59,7 @@ public interface Comment {
      *
      * @param col the 0-based column of the cell that contains the comment
      */
-    void setColumn(short col);
+    void setColumn(int col);
 
     /**
      * Name of the original comment author
